+++ src/Microsoft.PowerShell.Commands.Utility/commands/utility/ConvertFrom-SddlString.cs
@@ -215,7 +215,7 @@ namespace Microsoft.PowerShell.Commands
     /// <summary>
     /// Representation of a security descriptor.
     /// </summary>
-    public class SecurityDescriptorInfo
+    public sealed class SecurityDescriptorInfo
     {
         internal SecurityDescriptorInfo(
             string owner,
