+++ src/System.Management.Automation/engine/ExperimentalFeature/ExperimentalFeature.cs
@@ -339,75 +339,4 @@ namespace System.Management.Automation
         }
         private ExperimentAction _effectiveAction = ExperimentAction.None;
     }
-
-    /// <summary>
-    /// A readonly Hashset.
-    /// </summary>
-    internal sealed class ReadOnlyHashSet<T> : ISet<T>
-    {
-        private HashSet<T> _hashset;
-
-        /// <summary>
-        /// Constructor for the readonly Hashset.
-        /// </summary>
-        internal ReadOnlyHashSet(HashSet<T> hashset)
-        {
-            if (hashset == null)
-            {
-                throw new ArgumentNullException(nameof(hashset));
-            }
-            _hashset = hashset;
-        }
-
-        /// <summary>
-        /// Get an empty singleton.
-        /// </summary>
-        internal static readonly ReadOnlyHashSet<T> Empty = new ReadOnlyHashSet<T>(new HashSet<T>(capacity: 0));
-
-        /// <summary>
-        /// Get the count of the Hashset.
-        /// </summary>
-        public int Count => _hashset.Count;
-
-        /// <summary>
-        /// Indicate if it's a readonly Hashset.
-        /// </summary>
-        public bool IsReadOnly => true;
-
-        /// <summary>
-        /// Check if the set contains an item.
-        /// </summary>
-        public bool Contains(T item) => _hashset.Contains(item);
-
-        /// <summary>
-        /// Copy items to an array.
-        /// </summary>
-        public void CopyTo(T[] array, int arrayIndex) => _hashset.CopyTo(array, arrayIndex);
-
-        /// <summary>
-        /// GetEnumerator method.
-        /// </summary>
-        public IEnumerator<T> GetEnumerator() => _hashset.GetEnumerator();
-
-        #region NonPublic Interface Member Implementation
-
-        bool ISet<T>.IsProperSubsetOf(IEnumerable<T> other) => _hashset.IsProperSubsetOf(other);
-        bool ISet<T>.IsProperSupersetOf(IEnumerable<T> other) => _hashset.IsProperSupersetOf(other);
-        bool ISet<T>.IsSubsetOf(IEnumerable<T> other) => _hashset.IsSubsetOf(other);
-        bool ISet<T>.IsSupersetOf(IEnumerable<T> other) => _hashset.IsSupersetOf(other);
-        bool ISet<T>.Overlaps(IEnumerable<T> other) => _hashset.Overlaps(other);
-        bool ISet<T>.SetEquals(IEnumerable<T> other) => _hashset.SetEquals(other);
-        IEnumerator IEnumerable.GetEnumerator() => _hashset.GetEnumerator();
-
-        bool ISet<T>.Add(T item) => throw new NotSupportedException();
-        void ISet<T>.ExceptWith(IEnumerable<T> other) => throw new NotSupportedException();
-        void ISet<T>.IntersectWith(IEnumerable<T> other) => throw new NotSupportedException();
-        void ISet<T>.SymmetricExceptWith(IEnumerable<T> other) => throw new NotSupportedException();
-        void ISet<T>.UnionWith(IEnumerable<T> other) => throw new NotSupportedException();
-        void ICollection<T>.Add(T item) => throw new NotSupportedException();
-        void ICollection<T>.Clear() => throw new NotSupportedException();
-        bool ICollection<T>.Remove(T item) => throw new NotSupportedException();
-
-        #endregion
-    }
 }
